Fix degree 108 crashing bug
If x = Delta(s) = c_1[x_1] + higher terms where previously
Delta(t) = c'_1[x_1] + other higher terms = etr->full_cycle
(i.e. x_1 is a cycle in the table (etr->cycle) and is_member(start) is true),
this function is supposed to cancel the x_1 part of x with an
appropriate multiple of etr, so that now all terms of x are higher
than x_1. The correct multiple of etr to add to x is -c_1/c'_1, not
-c_1. To see that c'_1 is invertible, look at

    if(ModOper->ringOper->invertible(x[start])) return start;
    else return symplify_to_led(x,start+1,homotopy);

i.e. the only full_cycle's in the table have invertible coefficients for
the leading term.